requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
defusedxml==0.7.1
fastapi==0.116.1
fastapi-debug-toolbar==0.6.3
fastapi-pagination==0.14.1
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyinstrument==5.1.1
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
youtube-transcript-api==1.2.2
